Move ParserCompiler to separated class from Workflow
